html/semantics/embedded-content/the-img-element/update-the-image-data/element-delay-the-load-event.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Inline image element blocks load</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<img src="/images/blue.png?pipe=trickle(d7)">
<script>
var img_loaded = false;

var f = function() {
  img_loaded = true;
};

var img = document.querySelector("img");
if (img.complete) {
  f();
} else {
  img.onload = f;
}

async_test(function(t) {
  document.addEventListener("DOMContentLoaded", t.step_func_done(function() {
    assert_false(img_loaded);
  }));
}, "DOMContentLoaded doesn't wait for images");

async_test(function(t) {
  window.addEventListener("load", t.step_func_done(function() {
    assert_true(img_loaded);
  }));
}, "load waits for images");
</script>
